dom/base/test/test_bug1043106.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1043106
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1043106</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1043106">Mozilla Bug 1043106</a>
  <iframe id="iframe"></iframe>
  <script type="application/javascript">

var storage;

window.addEventListener("storage", function (event) {
  ok(event.storageArea, storage, "The storageArea is correct");
  storage.removeItem("a");
  runTests();
});

var tests = [ { key: 'localStorage', storage: localStorage },
              { key: 'sessionStorage', storage: sessionStorage } ];
function runTests() {
  if (!tests.length) {
    SimpleTest.finish();
    return;
  }

  var t = tests.shift();
  storage = t.storage;

  var ifr = document.getElementById("iframe");
  ifr.srcdoc = "<script>"+ t.key + ".setItem(\"a\",\"b\");</"+"script>";
}

SimpleTest.waitForExplicitFinish();
runTests();
  </script>
</body>
</html>
